refactor(test): split triggers/tests under cap + fix mismatch header
Split the 961-line triggers/tests.rs (over the 700-line hard cap) into a
triggers/tests/ directory along test-theme seams: trigger_type.rs for the
trigger_type() discriminant tables, and matching_{entity,mission,cover,misc}.rs
for the Trigger::matches() coverage. tests.rs becomes tests/mod.rs holding the
shared make_event fixture and mod decls. Pure move: all 58 tests preserved.

Also correct a misleading section header that labeled a mission-id mismatch
test as a step mismatch.
